+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Texture coordinates in geometries.
+- Texture coordinates projection based on geometry scope.
 
 ### Changed